Split parser tests by responsibility
