+++ crates/truthdb-sql/src/parser/mod.rs
@@ -3996,330 +3996,4 @@ fn is_reserved(keyword: &str) -> bool {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn alter_database_set_recovery_parses() {
-        let full = Parser::parse_str("ALTER DATABASE CURRENT SET RECOVERY FULL").expect("parse");
-        assert!(matches!(
-            &full[0],
-            Statement::AlterDatabase(a) if a.options == vec![(DatabaseOption::Recovery, true)]
-        ));
-        let simple = Parser::parse_str("ALTER DATABASE d SET RECOVERY SIMPLE").expect("parse");
-        assert!(matches!(
-            &simple[0],
-            Statement::AlterDatabase(a) if a.options == vec![(DatabaseOption::Recovery, false)]
-        ));
-        // Only FULL/SIMPLE are supported; any other mode is a syntax error.
-        assert!(Parser::parse_str("ALTER DATABASE d SET RECOVERY BULK_LOGGED").is_err());
-        // The existing ON/OFF options still parse alongside.
-        let rcsi =
-            Parser::parse_str("ALTER DATABASE d SET READ_COMMITTED_SNAPSHOT ON").expect("parse");
-        assert!(matches!(
-            &rcsi[0],
-            Statement::AlterDatabase(a)
-                if a.options == vec![(DatabaseOption::ReadCommittedSnapshot, true)]
-        ));
-    }
-
-    #[test]
-    fn backup_database_parses_with_options() {
-        let stmts = Parser::parse_str(
-            "BACKUP DATABASE mydb TO DISK = '/tmp/f.bak' WITH CHECKSUM, COPY_ONLY",
-        )
-        .expect("parse");
-        assert_eq!(stmts.len(), 1);
-        match &stmts[0] {
-            Statement::BackupDatabase {
-                database,
-                path,
-                checksum,
-                copy_only,
-                ..
-            } => {
-                assert_eq!(database.value, "mydb");
-                assert_eq!(path, "/tmp/f.bak");
-                assert!(*checksum);
-                assert!(*copy_only);
-            }
-            other => panic!("expected BackupDatabase, got {other:?}"),
-        }
-
-        // A bare BACKUP defaults CHECKSUM on and COPY_ONLY off.
-        let bare = Parser::parse_str("BACKUP DATABASE d TO DISK = 'x'").expect("parse bare");
-        assert!(matches!(
-            &bare[0],
-            Statement::BackupDatabase {
-                checksum: true,
-                copy_only: false,
-                ..
-            }
-        ));
-
-        // NO_CHECKSUM turns verification off; INIT is accepted (inert).
-        let no_ck = Parser::parse_str("BACKUP DATABASE d TO DISK = 'x' WITH NO_CHECKSUM, INIT")
-            .expect("parse");
-        assert!(matches!(
-            &no_ck[0],
-            Statement::BackupDatabase {
-                checksum: false,
-                ..
-            }
-        ));
-    }
-
-    #[test]
-    fn deeply_nested_parens_error_not_overflow() {
-        let sql = format!("SELECT {}1{}", "(".repeat(5000), ")".repeat(5000));
-        let err = Parser::parse_str(&sql).expect_err("must reject, not overflow");
-        assert_eq!(err.number, 191);
-    }
-
-    #[test]
-    fn deeply_nested_from_error_not_overflow() {
-        // Nested parenthesized-group FROM: must reject cleanly, not overflow.
-        let group = format!("SELECT 1 FROM {}t{}", "(".repeat(5000), ")".repeat(5000));
-        assert_eq!(Parser::parse_str(&group).unwrap_err().number, 191);
-        // Nested derived tables likewise.
-        let derived = format!(
-            "SELECT * FROM {}SELECT * FROM t{} x",
-            "(SELECT * FROM ".repeat(2000),
-            ") y".repeat(2000),
-        );
-        assert_eq!(Parser::parse_str(&derived).unwrap_err().number, 191);
-    }
-
-    #[test]
-    fn deep_not_and_unary_chains_error_not_overflow() {
-        let nots = format!("SELECT {}1", "NOT ".repeat(5000));
-        assert_eq!(Parser::parse_str(&nots).unwrap_err().number, 191);
-        // Spaced so `--` is not read as a comment.
-        let neg = format!("SELECT {}1", "- ".repeat(5000));
-        assert_eq!(Parser::parse_str(&neg).unwrap_err().number, 191);
-    }
-
-    #[test]
-    fn long_operator_chain_errors_not_overflow() {
-        // Parses iteratively but would overflow eval; the node budget caps it.
-        let sql = format!("SELECT 1{}", " OR 1".repeat(20_000));
-        assert_eq!(Parser::parse_str(&sql).unwrap_err().number, 191);
-    }
-
-    #[test]
-    fn the_node_budget_is_per_expression_not_per_batch() {
-        // Thousands of tiny FLAT expressions are fine — none deepens an
-        // evaluation spine. A per-batch count once made a 1001-tuple INSERT
-        // unparseable, which put row-lock escalation above the reachable
-        // ceiling.
-        let tuples: Vec<String> = (0..3000).map(|i| format!("({i}, {i})")).collect();
-        let sql = format!("INSERT INTO t VALUES {}", tuples.join(", "));
-        assert!(Parser::parse_str(&sql).is_ok());
-        // Two statements each near the budget: legal — the budget resets.
-        // Each `OR` costs ~2 nodes (the operator and its literal): 900 ORs
-        // ≈ 1801 nodes, inside the 2000 budget — twice over in one batch.
-        let chain = format!("SELECT 1{}", " OR 1".repeat(900));
-        let batch = format!("{chain}; {chain}");
-        assert!(Parser::parse_str(&batch).is_ok());
-        // One expression over the budget still errors, even at the end of an
-        // otherwise-light batch (1001 ORs ≈ 2003 nodes).
-        let over = format!("SELECT 1; SELECT 1{}", " OR 1".repeat(1001));
-        assert_eq!(Parser::parse_str(&over).unwrap_err().number, 191);
-        // CTE and derived-table bodies parse under depth >= 1 (no depth-0
-        // reset for their expressions): the per-statement reset must cover
-        // them, or a big-but-legal statement poisons the next one's budget.
-        let big = format!("SELECT 1{}", " OR 1".repeat(900));
-        let derived = format!("{big}; SELECT * FROM (SELECT 1{}) d", " OR 1".repeat(200));
-        assert!(Parser::parse_str(&derived).is_ok(), "derived after big");
-        let cte = format!(
-            "{big}; WITH c AS (SELECT 1{} AS x) SELECT x FROM c",
-            " OR 1".repeat(200)
-        );
-        assert!(Parser::parse_str(&cte).is_ok(), "cte after big");
-    }
-
-    #[test]
-    fn reasonable_depth_is_accepted() {
-        let sql = format!("SELECT {}1{}", "(".repeat(50), ")".repeat(50));
-        assert!(Parser::parse_str(&sql).is_ok());
-        let chain = format!("SELECT 1{}", " + 1".repeat(100));
-        assert!(Parser::parse_str(&chain).is_ok());
-    }
-
-    #[test]
-    fn try_catch_parses_into_blocks() {
-        let stmts = Parser::parse_str(
-            "BEGIN TRY \
-               INSERT INTO t VALUES (1); \
-               SELECT 2; \
-             END TRY \
-             BEGIN CATCH \
-               SELECT ERROR_NUMBER(); \
-             END CATCH",
-        )
-        .expect("parse");
-        let Statement::TryCatch {
-            try_block,
-            catch_block,
-            ..
-        } = &stmts[0]
-        else {
-            panic!("expected a TRY/CATCH, got {:?}", stmts[0]);
-        };
-        assert_eq!(try_block.len(), 2, "two statements in the TRY block");
-        assert_eq!(catch_block.len(), 1, "one statement in the CATCH block");
-        assert!(matches!(try_block[0], Statement::Insert(_)));
-
-        // A nested TRY inside the TRY block is consumed whole (its END TRY / END
-        // CATCH do not close the outer block).
-        let stmts = Parser::parse_str(
-            "BEGIN TRY \
-               BEGIN TRY SELECT 1; END TRY BEGIN CATCH SELECT 2; END CATCH; \
-               SELECT 3; \
-             END TRY \
-             BEGIN CATCH SELECT 4; END CATCH",
-        )
-        .expect("parse");
-        let Statement::TryCatch { try_block, .. } = &stmts[0] else {
-            panic!("expected a TRY/CATCH");
-        };
-        assert_eq!(try_block.len(), 2, "nested TRY + the following SELECT");
-        assert!(matches!(try_block[0], Statement::TryCatch { .. }));
-
-        // An unterminated TRY block is a syntax error, not a hang.
-        assert_eq!(
-            Parser::parse_str("BEGIN TRY SELECT 1;").unwrap_err().number,
-            102,
-        );
-    }
-
-    #[test]
-    fn try_catch_parses_without_statement_terminators() {
-        // The canonical T-SQL form omits the `;` before END TRY / END CATCH.
-        // `END` must not be read as an implicit alias for the preceding select
-        // item (or table), which would leave the cursor on `TRY`.
-        for sql in [
-            "BEGIN TRY SELECT 1 END TRY BEGIN CATCH SELECT 2 END CATCH",
-            "BEGIN TRY SELECT * FROM t END TRY BEGIN CATCH SELECT 2 END CATCH",
-            "BEGIN TRY SELECT a FROM t WHERE a = 1 END TRY \
-             BEGIN CATCH SELECT ERROR_MESSAGE() END CATCH",
-        ] {
-            let stmts = Parser::parse_str(sql).unwrap_or_else(|e| panic!("{sql}: {e:?}"));
-            let Statement::TryCatch {
-                try_block,
-                catch_block,
-                ..
-            } = &stmts[0]
-            else {
-                panic!("expected a TRY/CATCH for {sql}");
-            };
-            assert_eq!(try_block.len(), 1, "{sql}");
-            assert_eq!(catch_block.len(), 1, "{sql}");
-        }
-
-        // An explicit `AS end` still aliases (only the *bare* END is declined),
-        // and a delimited [end] is an identifier, not the block terminator.
-        let stmts = Parser::parse_str("SELECT 1 AS end").expect("AS end still aliases");
-        assert!(matches!(stmts[0], Statement::Select(_)));
-        let stmts = Parser::parse_str("SELECT 1 [end]").expect("[end] still aliases");
-        assert!(matches!(stmts[0], Statement::Select(_)));
-    }
-
-    #[test]
-    fn assignment_select_parses_as_assign_item() {
-        // `SELECT @v = expr` is an assignment item, not a boolean comparison.
-        let stmts = Parser::parse_str("SELECT @v = 1 + 2").expect("parse");
-        let Statement::Select(select) = &stmts[0] else {
-            panic!("expected select")
-        };
-        assert!(
-            matches!(&select.items[0], SelectItem::Assign { target, .. } if target == "v"),
-            "expected an assignment item: {:?}",
-            select.items[0]
-        );
-
-        // `@v = x` inside a WHERE stays a comparison (only the item list assigns).
-        let stmts = Parser::parse_str("SELECT 1 WHERE @v = 5").expect("parse");
-        let Statement::Select(select) = &stmts[0] else {
-            panic!("expected select")
-        };
-        assert!(matches!(&select.items[0], SelectItem::Expr { .. }));
-        assert!(select.where_clause.is_some());
-
-        // Mixing an assignment with a result column is a syntax-level error 141.
-        assert_eq!(
-            Parser::parse_str("SELECT @v = 1, 2").unwrap_err().number,
-            141,
-        );
-    }
-
-    #[test]
-    fn ignorable_set_options_parse_as_noops() {
-        // Cosmetic/advisory options clients send at connection time: ON/OFF
-        // flags, value forms, a signed value, and a required-ON option at ON.
-        // (NOCOUNT graduated to a real option in Stage 14.)
-        let sql = "SET QUOTED_IDENTIFIER ON; SET ANSI_WARNINGS OFF; \
-                   SET TEXTSIZE 2147483647; SET DATEFORMAT mdy; SET LOCK_TIMEOUT -1";
-        let stmts = Parser::parse_str(sql).expect("all recognized as no-ops");
-        assert_eq!(stmts.len(), 5);
-        assert!(
-            stmts
-                .iter()
-                .all(|s| matches!(s, Statement::Set(SetStatement::Ignored))),
-            "every option should parse to SetStatement::Ignored: {stmts:?}",
-        );
-        // NOCOUNT is a real session option now.
-        let stmts = Parser::parse_str("SET NOCOUNT ON; SET NOCOUNT OFF").expect("parses");
-        assert!(matches!(
-            stmts.as_slice(),
-            [
-                Statement::Set(SetStatement::NoCount(true)),
-                Statement::Set(SetStatement::NoCount(false))
-            ]
-        ));
-        // An unknown option is still a syntax error, not silently ignored.
-        assert_eq!(Parser::parse_str("SET WHATSIT ON").unwrap_err().number, 102);
-    }
-
-    #[test]
-    fn result_changing_set_options_are_not_silently_ignored() {
-        // OFF for an option TruthDB hardwires to ON must be rejected, never
-        // silently accepted (it would change query results).
-        assert_eq!(
-            Parser::parse_str("SET ANSI_NULLS OFF").unwrap_err().number,
-            102,
-        );
-        assert_eq!(
-            Parser::parse_str("SET CONCAT_NULL_YIELDS_NULL OFF")
-                .unwrap_err()
-                .number,
-            102,
-        );
-        // ...but the matching ON is a faithful no-op.
-        assert!(matches!(
-            Parser::parse_str("SET ANSI_NULLS ON").as_deref(),
-            Ok([Statement::Set(SetStatement::Ignored)]),
-        ));
-        // Options that change what/how much runs stay hard errors, not no-ops,
-        // so we never silently drop a client's row cap or skip flag.
-        for sql in [
-            "SET ROWCOUNT 100",
-            "SET NOEXEC ON",
-            "SET IMPLICIT_TRANSACTIONS ON",
-        ] {
-            assert_eq!(
-                Parser::parse_str(sql).unwrap_err().number,
-                102,
-                "{sql} must not be a silent no-op",
-            );
-        }
-    }
-
-    #[test]
-    fn table_level_primary_key_duplicate_rejected() {
-        let sql = "CREATE TABLE t (a INT, PRIMARY KEY (a), PRIMARY KEY (a))";
-        assert_eq!(Parser::parse_str(sql).unwrap_err().number, 8110);
-        let sql2 = "CREATE TABLE t (id INT PRIMARY KEY, PRIMARY KEY (id))";
-        assert_eq!(Parser::parse_str(sql2).unwrap_err().number, 8110);
-    }
-}
+mod tests;
