+++ crates/truthdb-sql/src/parser/tests/mod.rs
@@ -1,0 +1,7 @@
+use super::*;
+
+mod control;
+mod ddl;
+mod depth;
+mod query;
+mod session;
